First prototype for autocomplete - see description
The search terms are in an array on lines 90-92 in main.js. There are multiple versions that I tried to get working. Apparently, the search_terms array has to have elements that are separated by single quotes ' ', not double quotes " ", for it to work. Not really sure why. I tried splitting the input file into an array that splits by both newlines and spaces, but could not come up with a surefire solution (hence, why line 92 is commented out right now). The array in line 90 is the entire text file separated correctly, but only by the full phrase. It is also hardcoded, which we probably can't do.